dev_requirements.txt
requests
simplejson
pandas
Shapely>=1.7.0
geopandas>=0.8.1
djangorestframework<3.13
Django<=4.0.0

https://github.com/Murabei-OpenSource-Codes/pumpwood-communication/archive/refs/tags/1.33.zip
https://github.com/Murabei-OpenSource-Codes/pumpwood-miscellaneous/archive/refs/tags/0.25.zip
https://github.com/Murabei-OpenSource-Codes/pumpwood-i8n/archive/refs/tags/0.1.4.zip
https://github.com/Murabei-OpenSource-Codes/pumpwood-djangoauth/archive/refs/tags/1.60.21.zip
https://github.com/Murabei-OpenSource-Codes/pumpwood-kong/archive/refs/tags/0.7.zip
pdoc
